refactor: 매서드 분리와 mapper 클래스 정의,import 경로 수정
