+++ src/main/java/kakao/rebit/book/controller/BookController.java
@@ -1,8 +1,8 @@
 package kakao.rebit.book.controller;
 
 import java.util.List;
+import kakao.rebit.book.dto.BookResponse;
 import kakao.rebit.book.service.BookService;
-import kakao.rebit.feed.dto.response.BookResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
